Rewrite the storage of external events
Previously, the external events were stored in dedicated slots
attached to each event buffer.  This created a number of problems:

* the number of slots was fixed and set across the different
  importance levels; as a result, we made rather inefficient use of
  that memory.

* the events could have been unregistered, but their slot was not
  freed until that range of event IDs was rotated out of the
  buffer. this behavior was rather counterintuitive, and did not leave
  the app developers with the control they wanted.

This patch replaces the dedicated slots for external events with
sentinel entries for external events in the regular event buffers.
The result is that the effective number of slots is bounded by the
event buffer sizes rather than additional configuration parameters,
and external events are flushed the same way as all other event
entries.  The existing API for external events is modified slightly:
instead of passing around pointers as handles to the external events,
we identify external event blocks by the last event ID in the
sequence.  The full structure of ExternalEvents remains accessible
within the FetchExternalEventsFunct as an additional member of the
EventLoadOutContext.

Change-Id: Ie42c622e7bf466e2ac803bc32c362bd17a98d532